// .backup/2026-10-02_15-45-50/new/src/types.ts
export type ClinicalSystemKey =
  | 'headache'
  | 'skin_hair'
  | 'gastrointestinal'
  | 'urinary'
  | 'musculoskeletal'
  | 'respiratory'
  | 'female_gynae'
  | 'pediatric'
  | 'other_mind_generals';
// ...
export type QuestionType =
  | 'short_text' | 'long_text' | 'number' | 'date' | 'time'
  | 'single_choice' | 'multiple_choice' | 'dropdown' | 'yes_no'
  | 'image_upload' | 'multiple_image_upload' | 'file_upload' | 'section_text';